refactor(cli): split apiCommand by concern + unify stdin-read helper
Two small follow-ups after PR-review feedback on the 1312-line
commands/api.ts:

1. Split the file into one-concern modules under commands/api/:
   - types.ts        (ApiCommandOptions, ApiCommandIO, DEFAULT_IO)
   - url.ts          (buildUrl + HostMismatchError)
   - headers.ts      (buildHeaders)
   - body.ts         (buildBody, -G projection, --data-urlencode)
   - method.ts       (pickMethod, isHttpMethod)
   - stream.ts       (consumeResponseStream, ChunkSink family, abortAsRejection)
   - retry.ts        (executeWithRetry + backoff helpers)
   - verbose.ts      (writeVerboseRequest/Response)
   - write-out.ts    (formatWriteOut, WriteOutMetrics, sizeOfBody)

   commands/api.ts keeps the orchestrator (~450 lines) and re-exports
   the public surface (apiCommand, isHttpMethod, ApiCommandOptions,
   ApiCommandIO, HostMismatchError, WriteOutMetrics) so cli.ts and
   the test suite import unchanged.

2. Unify the two inline copies of "read a body-data ref (path or
   stdin) into a string" that lived inside collectGetDataAsQuery and
   extractUrlencodeParts. They had already drifted (chunk accumulation
   shape, empty-reader early-exit); a single readRefAsString(ref, io)
   in body.ts is the one source of truth, so future fixes land once.

No behavior change: full CLI suite (472 tests) green, tsc clean,
prettier + eslint clean. Follows the lib/install/ precedent — the
orchestrator stays in commands/, the helpers live in commands/api/.
